Move classes to internal package
