+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/DefaultRoleSetupImportService.java
@@ -7,10 +7,10 @@ import uk.gov.hmcts.reform.amlib.enums.AccessType;
 import uk.gov.hmcts.reform.amlib.enums.RoleType;
 import uk.gov.hmcts.reform.amlib.enums.SecurityClassification;
 import uk.gov.hmcts.reform.amlib.exceptions.PersistenceException;
+import uk.gov.hmcts.reform.amlib.internal.repositories.DefaultRoleSetupRepository;
+import uk.gov.hmcts.reform.amlib.internal.repositories.models.ResourceAttribute;
+import uk.gov.hmcts.reform.amlib.internal.repositories.models.RoleBasedAccessRecord;
 import uk.gov.hmcts.reform.amlib.models.DefaultPermissionGrant;
-import uk.gov.hmcts.reform.amlib.models.ResourceAttribute;
-import uk.gov.hmcts.reform.amlib.models.RoleBasedAccessRecord;
-import uk.gov.hmcts.reform.amlib.repositories.DefaultRoleSetupRepository;
 
 import javax.sql.DataSource;
 
@@ -21,7 +21,7 @@ public class DefaultRoleSetupImportService {
      * This constructor has issues with performance due to requiring a new connection for every query.
      *
      * @param url      the url for the database
-     * @param username     the username for the database
+     * @param username the username for the database
      * @param password the password for the database
      */
     public DefaultRoleSetupImportService(String url, String username, String password) {
