+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/internal/repositories/AccessManagementRepository.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.amlib.repositories;
+package uk.gov.hmcts.reform.amlib.internal.repositories;
 
 import org.jdbi.v3.sqlobject.config.RegisterColumnMapper;
 import org.jdbi.v3.sqlobject.config.RegisterConstructorMapper;
@@ -6,11 +6,11 @@ import org.jdbi.v3.sqlobject.customizer.BindBean;
 import org.jdbi.v3.sqlobject.statement.SqlQuery;
 import org.jdbi.v3.sqlobject.statement.SqlUpdate;
 import uk.gov.hmcts.reform.amlib.enums.AccessType;
+import uk.gov.hmcts.reform.amlib.internal.repositories.mappers.JsonPointerMapper;
+import uk.gov.hmcts.reform.amlib.internal.repositories.mappers.PermissionSetMapper;
+import uk.gov.hmcts.reform.amlib.internal.repositories.models.ExplicitAccessRecord;
+import uk.gov.hmcts.reform.amlib.internal.repositories.models.RoleBasedAccessRecord;
 import uk.gov.hmcts.reform.amlib.models.ExplicitAccessMetadata;
-import uk.gov.hmcts.reform.amlib.models.ExplicitAccessRecord;
-import uk.gov.hmcts.reform.amlib.models.RoleBasedAccessRecord;
-import uk.gov.hmcts.reform.amlib.repositories.mappers.JsonPointerMapper;
-import uk.gov.hmcts.reform.amlib.repositories.mappers.PermissionSetMapper;
 
 import java.util.List;
