+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/internal/repositories/models/ExplicitAccessRecord.java
@@ -1,13 +1,20 @@
-package uk.gov.hmcts.reform.amlib.models;
+package uk.gov.hmcts.reform.amlib.internal.repositories.models;
 
 import com.fasterxml.jackson.core.JsonPointer;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import uk.gov.hmcts.reform.amlib.enums.Permission;
 import uk.gov.hmcts.reform.amlib.enums.SecurityClassification;
+import uk.gov.hmcts.reform.amlib.internal.utils.Permissions;
+import uk.gov.hmcts.reform.amlib.models.AttributeAccessDefinition;
+
+import java.util.Set;
 
 @Data
 @Builder
-public final class ExplicitAccessMetadata {
+@AllArgsConstructor
+public final class ExplicitAccessRecord implements AttributeAccessDefinition {
     private final String resourceId;
     private final String accessorId;
     private final String accessType;
@@ -15,9 +22,16 @@ public final class ExplicitAccessMetadata {
     private final String resourceType;
     private final String resourceName;
     private final JsonPointer attribute;
+    private final Set<Permission> permissions;
     private final SecurityClassification securityClassification;
 
+    @Override
     public String getAttributeAsString() {
         return getAttribute().toString();
+    }
+
+    @Override
+    public int getPermissionsAsInt() {
+        return Permissions.sumOf(permissions);
     }
 }
