+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/internal/repositories/models/ResourceDefinition.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.amlib.models;
+package uk.gov.hmcts.reform.amlib.internal.repositories.models;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 @Builder
 @AllArgsConstructor
-public class ResourceDefinition {
+public final class ResourceDefinition {
     private final String serviceName;
     private final String resourceType;
     private final String resourceName;
