+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/internal/repositories/models/RoleBasedAccessRecord.java
@@ -1,11 +1,12 @@
-package uk.gov.hmcts.reform.amlib.models;
+package uk.gov.hmcts.reform.amlib.internal.repositories.models;
 
 import com.fasterxml.jackson.core.JsonPointer;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import uk.gov.hmcts.reform.amlib.enums.Permission;
-import uk.gov.hmcts.reform.amlib.utils.Permissions;
+import uk.gov.hmcts.reform.amlib.internal.utils.Permissions;
+import uk.gov.hmcts.reform.amlib.models.AttributeAccessDefinition;
 
 import java.util.Set;
 
@@ -13,7 +14,6 @@ import java.util.Set;
 @Builder
 @AllArgsConstructor
 public final class RoleBasedAccessRecord implements AttributeAccessDefinition {
-
     private final String serviceName;
     private final String resourceType;
     private final String resourceName;
