+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/internal/utils/Permissions.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.amlib.utils;
+package uk.gov.hmcts.reform.amlib.internal.utils;
 
 import uk.gov.hmcts.reform.amlib.enums.Permission;
 import uk.gov.hmcts.reform.amlib.exceptions.UnsupportedPermissionsException;
