--- am-lib/src/main/java/uk/gov/hmcts/reform/amlib/models/AbstractAccessMetadata.java
@@ -1,22 +1,0 @@
-package uk.gov.hmcts.reform.amlib.models;
-
-import com.fasterxml.jackson.core.JsonPointer;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import uk.gov.hmcts.reform.amlib.enums.SecurityClassification;
-
-@Data
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-@SuppressWarnings("PMD")
-public abstract class AbstractAccessMetadata {
-
-    private final String resourceId;
-    private final String accessorId;
-    private final String accessType;
-    private final String serviceName;
-    private final String resourceType;
-    private final String resourceName;
-    private final JsonPointer attribute;
-    private final SecurityClassification securityClassification;
-}
