+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/models/DefaultPermissionGrant.java
@@ -1,20 +1,17 @@
 package uk.gov.hmcts.reform.amlib.models;
 
 import com.fasterxml.jackson.core.JsonPointer;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import uk.gov.hmcts.reform.amlib.enums.Permission;
 import uk.gov.hmcts.reform.amlib.enums.SecurityClassification;
-import uk.gov.hmcts.reform.amlib.utils.Pair;
 
 import java.util.Map;
 import java.util.Set;
 
 @Data
 @Builder
-@AllArgsConstructor
-public class DefaultPermissionGrant {
+public final class DefaultPermissionGrant {
     private final String serviceName;
     private final String resourceType;
     private final String resourceName;
