+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/models/ExplicitAccessGrant.java
@@ -1,7 +1,6 @@
 package uk.gov.hmcts.reform.amlib.models;
 
 import com.fasterxml.jackson.core.JsonPointer;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import uk.gov.hmcts.reform.amlib.enums.Permission;
@@ -12,9 +11,7 @@ import java.util.Set;
 
 @Data
 @Builder
-@AllArgsConstructor
-public class ExplicitAccessGrant {
-
+public final class ExplicitAccessGrant {
     private final String resourceId;
     private final String accessorId;
     private final String accessType;
