+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/models/Resource.java
@@ -3,10 +3,11 @@ package uk.gov.hmcts.reform.amlib.models;
 import com.fasterxml.jackson.databind.JsonNode;
 import lombok.Builder;
 import lombok.Data;
+import uk.gov.hmcts.reform.amlib.internal.repositories.models.ResourceDefinition;
 
 @Data
 @Builder
-public class Resource {
+public final class Resource {
     private final String resourceId;
     private final ResourceDefinition type;
     private final JsonNode resourceJson;
