+++ am-lib/src/test/java/integration/uk/gov/hmcts/reform/amlib/importer/RoleIntegrationTest.java
@@ -7,7 +7,7 @@ import uk.gov.hmcts.reform.amlib.DefaultRoleSetupImportService;
 import uk.gov.hmcts.reform.amlib.enums.AccessType;
 import uk.gov.hmcts.reform.amlib.enums.RoleType;
 import uk.gov.hmcts.reform.amlib.enums.SecurityClassification;
-import uk.gov.hmcts.reform.amlib.models.Role;
+import uk.gov.hmcts.reform.amlib.internal.repositories.models.Role;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static uk.gov.hmcts.reform.amlib.helpers.TestConstants.ROLE_NAME;
